requirements.txt
redis
requests
PyExecJS